Fix "Defer cleanup of old addressbook accounts"
This change was lost when rebasing peoplesync-3.4

Fixes 03c58225c6880bb7d8f020808cf7fb9711f57831.

+++ app/src/main/java/com/messageconcept/peoplesyncclient/syncadapter/AccountUtils.kt
@@ -44,7 +44,7 @@ object AccountUtils {
             .map { LocalAddressBook(context, it, null) }
             .forEach {
                 try {
-                    if (!accountNames.contains(it.mainAccount.name))
+                    if (!accountNames.contains(it.mainAccount.name) && newAccountAvailable)
                         it.delete()
                 } catch(e: Exception) {
                     Logger.log.log(Level.SEVERE, "Couldn't delete address book account", e)
